1.0.9 update
Added some fields which can be used to aid accessibility of the rating widget. There is an `ARIA Labelledby value` field which can be used to associate the rating widget with an HTML element. There are two `additional assistive text` fields can be used to add text which is only `visible` to screen-reader users, these will add span elements with a class of `show-for-sr` to the default `a` anchor button type or will add an `aria-label` with a combination of the regular label and the additional assistive text label to the `button` type, note that you will need to add appropriate CSS styles to render these as hidden to visual users but accessible by screen-reader users.